apps/mcp/pyproject.toml
[project]
name = "e2b-sandbox-mcp-server"
version = "0.1.0"
description = "MCP server that wraps the E2B Sandbox CLI for agent sandbox access"
readme = "README.md"
requires-python = ">=3.12"

dependencies = [
    "sandbox-common",
    "mcp>=1.0.0",
    "pytest>=8.0.0",
    "python-dotenv>=1.0.0",
]

[tool.uv.sources]
sandbox-common = { path = "../../lib", editable = true }
